requirements.txt
attrs==23.2.0
bcrypt==4.1.2
bleach==6.1.0
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
click==8.1.7
cryptography==42.0.7
defusedxml==0.7.1
dnspython==2.6.1
email_validator==2.1.1
Flask==3.0.2
Flask-Login==0.6.3
flask-mailman==1.0.0
Flask-Principal==0.4.0
flask-saml2-fork @ git+https://github.com/WRLC/flask-saml2.git@5ec183edf8edf3b6a9119f67f917df2d8f78867c
Flask-Security-Too==5.4.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.3
gunicorn==22.0.0
idna==3.6
importlib_resources==6.4.0
iso8601==0.1.16
itsdangerous==2.1.2
Jinja2==3.1.3
lxml==4.9.4
MarkupSafe==2.1.5
packaging==24.0
passlib==1.7.4
pycparser==2.22
PyMySQL==1.1.0
pyOpenSSL==24.1.0
python-dotenv==1.0.1
pytz==2024.1
signxml==3.2.1
six==1.16.0
SQLAlchemy==2.0.29
SQLAlchemy-Utils==0.41.2
typing_extensions==4.11.0
webencodings==0.5.1
Werkzeug==3.0.2
WTForms==3.1.2
